feat(dap): split cpp configuration
